+++ projects/plugins/jetpack/tools/webpack.config.css.js
@@ -55,7 +55,7 @@ const sharedWebpackConfig = {
 						options: {
 							sassOptions: {
 								// The minifier will minify if necessary.
-								outputStyle: 'expanded',
+								style: 'expanded',
 							},
 						},
 					},