+++ projects/js-packages/shared-extension-utils/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.19.3] - 2025-05-12
+### Changed
+- Update dependencies. [#43068]
+
 ## [0.19.2] - 2025-05-05
 ### Changed
 - Update package dependencies. [#43320] [#43326]
@@ -621,6 +625,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Core: prepare utility for release
 
+[0.19.3]: https://github.com/Automattic/jetpack-shared-extension-utils/compare/0.19.2...0.19.3
 [0.19.2]: https://github.com/Automattic/jetpack-shared-extension-utils/compare/0.19.1...0.19.2
 [0.19.1]: https://github.com/Automattic/jetpack-shared-extension-utils/compare/0.19.0...0.19.1
 [0.19.0]: https://github.com/Automattic/jetpack-shared-extension-utils/compare/0.18.11...0.19.0