+++ projects/packages/jitm/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.2.16] - 2025-05-12
+### Changed
+- Update package dependencies. [#43400]
+
 ## [4.2.15] - 2025-05-05
 ### Changed
 - Update package dependencies. [#43326]
@@ -905,6 +909,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Update Jetpack to use new JITM package
 
+[4.2.16]: https://github.com/Automattic/jetpack-jitm/compare/v4.2.15...v4.2.16
 [4.2.15]: https://github.com/Automattic/jetpack-jitm/compare/v4.2.14...v4.2.15
 [4.2.14]: https://github.com/Automattic/jetpack-jitm/compare/v4.2.13...v4.2.14
 [4.2.13]: https://github.com/Automattic/jetpack-jitm/compare/v4.2.12...v4.2.13